Security

+++ src/main/java/fit/health/fithealthapi/controllers/UserController.java
@@ -5,13 +5,19 @@ import fit.health.fithealthapi.exceptions.UserNotFoundException;
 import fit.health.fithealthapi.model.Recipe;
 import fit.health.fithealthapi.model.User;
 import fit.health.fithealthapi.model.dto.EditUserDTO;
-import fit.health.fithealthapi.model.dto.UserDTO;
+import fit.health.fithealthapi.model.enums.Role;
 import fit.health.fithealthapi.services.RecipeService;
 import fit.health.fithealthapi.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/users")
@@ -22,21 +28,20 @@ public class UserController {
     @Autowired
     private RecipeService recipeService;
 
-    @PostMapping("/register")
-    public ResponseEntity<?> createUser(@RequestBody UserDTO user) {
-        try {
-            User createdUser = userService.saveUser(user);
-            return ResponseEntity.status(HttpStatus.CREATED).body(createdUser);
-        } catch (IllegalArgumentException e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid user data: " + e.getMessage());
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred: " + e.getMessage());
-        }
+    @PostMapping("/logout")
+    public ResponseEntity<?> logout() {
+        SecurityContextHolder.clearContext(); // Clears the user's security context
+        return ResponseEntity.ok("Logged out successfully");
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateUser(@PathVariable Long id, @RequestBody EditUserDTO user) {
+    public ResponseEntity<?> updateUser(@PathVariable Long id, @RequestBody EditUserDTO user, Authentication authentication) {
         try {
+            String currentUsername = authentication.getName();
+            User currentUser = userService.getUserByUsername(currentUsername);
+            if (!currentUser.getRole().equals(Role.ADMIN) && !currentUser.getId().equals(id)) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You don't have permission to edit this user.");
+            }
             User updatedUser = userService.updateUser(id, user);
             return ResponseEntity.ok(updatedUser);
         } catch (UserNotFoundException e) {
@@ -48,21 +53,14 @@ public class UserController {
         }
     }
 
-    @PostMapping("/login")
-    public ResponseEntity<?> login(@RequestBody UserDTO user) {
-        try {
-            User loggedInUser = userService.checkCredentials(user.getUsername(), user.getPassword());
-            return ResponseEntity.ok(loggedInUser);
-        } catch (IllegalArgumentException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid username or password.");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred: " + e.getMessage());
-        }
-    }
-
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<?> deleteUser(@PathVariable Long id, Authentication authentication) {
         try {
+            String currentUsername = authentication.getName();
+            User currentUser = userService.getUserByUsername(currentUsername);
+            if (!currentUser.getRole().equals(Role.ADMIN) && !currentUser.getId().equals(id)) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You don't have permission to delete this user.");
+            }
             userService.deleteUser(id);
             return ResponseEntity.noContent().build();
         } catch (UserNotFoundException e) {
